rename Injection to CboxApplication
