+++ lib/blocks/inc/OneWireScanningFactory.hpp
@@ -23,8 +23,8 @@
 #include "blocks/DS2413Block.h"
 #include "blocks/TempSensorOneWireBlock.h"
 #include "blox_hal/hal_delay.h"
+#include "cbox/CboxApplication.h"
 #include "cbox/CboxPtr.h"
-#include "cbox/Injection.h"
 #include "cbox/Object.h"
 #include "cbox/ObjectContainer.h"
 #include "cbox/ScanningFactory.hpp"
