+++ lib/cbox/inc/cbox/ScanningFactory.hpp
@@ -20,7 +20,7 @@
 
 #pragma once
 
-#include "cbox/Injection.h"
+#include "cbox/CboxApplication.h"
 #include "cbox/Object.h"
 #include "cbox/ObjectContainer.h"
 #include <memory>
