+++ test/blocks/TestCboxApplication.cpp
@@ -1,12 +1,9 @@
 #include "OneWireScanningFactory.hpp"
-#include "SparkEepromAccess.h"
-#include "blocks/OneWireBusBlock.h"
 #include "brewblox.hpp"
-#include "cbox/CboxPtr.h"
+#include "cbox/ArrayEepromAccess.h"
+#include "cbox/CboxApplication.h"
 #include "cbox/EepromObjectStorage.h"
-#include "cbox/Injection.h"
 #include "cbox/ObjectFactory.h"
-#include "control/OneWire.h"
 
 namespace cbox {
 
@@ -25,8 +22,8 @@ std::shared_ptr<Object> scan()
 
 ObjectStorage& getStorage()
 {
-    static SparkEepromAccess eeprom;
-    static EepromObjectStorage objectStore(eeprom);
+    static cbox::ArrayEepromAccess<2048> eeprom;
+    static cbox::EepromObjectStorage objectStore(eeprom);
     return objectStore;
 }
